+++ .pre-commit-config.yaml
@@ -416,8 +416,8 @@ repos:
         name: Generate airflow diagrams
         entry: ./scripts/ci/pre_commit/pre_commit_generate_airflow_diagrams.py
         language: python
-        files: ^scripts/ci/pre_commit/pre_commit_generate_airflow_diagrams.py
-        pass_filenames: false
+        files: ^docs/.*/diagram_[^/]*\.py$
+        pass_filenames: true
         additional_dependencies: ['rich>=12.4.4', "diagrams>=0.23.4"]
       - id: update-supported-versions
         name: Updates supported versions in documentation
