+++ docs/apache-airflow-providers-fab/img/diagram_fab_auth_manager_airflow_architecture.md5sum
@@ -1,0 +1,1 @@
+aa73a8292341145e0f60682f7047503b
