+++ docs/apache-airflow/img/diagram_auth_manager_airflow_architecture.md5sum
@@ -1,0 +1,1 @@
+5b82cba489898a46dcfe5f458eeee33b
