+++ docs/diagrams/python_multiprocess_logo.png
